Section 2, Lecture 49: Make a Pin by Spin
Learnt what is Spin

Deleted the cube and rename the Curves

Convert the Curves into mesh under mesh tab (alt with c), Mesh from Curves
In left side properties, tick keep original will save your curves as well. Rename Both.

In Edit Mode, Select All (a), in tools menu click Spin. It similar as Extrude (e)

Set the Center As 0,0,0
Set the Axis 0,0,?

Center - Moves the models central point of which it will spin around
Axis - The axis is in which direction the rotation will be around that center point.