Section 3, Lecture 54: Make Hole of Bowling Ball
Learn and how to make a hole(Sphere shape) into Bowling Balls.

Use Subdivide(w) of the Edges Loop(alt with click(edge)),
and To Sphere (shift with alt with s)

To do this in Ico Sphere, select the hexagon (six triangle), Insert it (i), the reason to insert because to do in edges loop. After that, Select the edges loop and subdivides to suitable value such as 3.
Use To Sphere to make it in the sphere shape.