Section 3, Lecture 55: Make Hole to BowlingBall
In Previous, we need to make a hole after To Shpere we need to Insert(i) Once again for the outsider more like Sphere in depth of hole.

Select the new Face after Insert.

Use Extrude (e) however the option in left hand side properties we need the orientation is Nomral. It Z axis will follow the vertices of center (try it out).
Extrude in Negative value to indepth.

Select individual hole to do to avoid the Normal center point are different

Add new materials to the Ball